+++ bigtangle-server/src/test/java/net/bigtangle/server/TokenDomainNameTest.java
@@ -1,5 +1,6 @@
 package net.bigtangle.server;
 
+import java.math.BigInteger;
 import java.util.ArrayList;
 import java.util.List;
 
@@ -29,8 +30,8 @@ public class TokenDomainNameTest extends AbstractIntegrationTest {
 
         {
             final String tokenid = walletKeys.get(0).getPublicKeyAsHex();
-            walletAppKit1.wallet().publishDomainName(walletKeys.get(0), tokenid, "bigtangle.bc",
-                    "bigtangle.bc", aesKey, 678900000, "");
+            walletAppKit1.wallet().publishDomainName(walletKeys.get(0), tokenid, "bigtangle.bc", "bigtangle.bc", aesKey,
+                    BigInteger.valueOf(678900000), "");
 
             List<ECKey> keys = new ArrayList<ECKey>();
             keys.add(preKey);
@@ -41,8 +42,8 @@ public class TokenDomainNameTest extends AbstractIntegrationTest {
 
         {
             final String tokenid = walletKeys.get(1).getPublicKeyAsHex();
-            walletAppKit1.wallet().publishDomainName(walletKeys.get(1), tokenid, "www.bigtangle.bc",
-                    "www.bigtangle.bc", aesKey, 678900000, "");
+            walletAppKit1.wallet().publishDomainName(walletKeys.get(1), tokenid, "www.bigtangle.bc", "www.bigtangle.bc",
+                    aesKey, BigInteger.valueOf(678900000), "");
             walletAppKit1.wallet().multiSign(tokenid, preKey, aesKey);
 
             List<ECKey> keys = new ArrayList<ECKey>();
@@ -58,8 +59,8 @@ public class TokenDomainNameTest extends AbstractIntegrationTest {
         {
             final String tokenid = walletKeys.get(2).getPublicKeyAsHex();
             walletAppKit1.wallet().publishDomainName(ImmutableList.of(walletKeys.get(2), walletKeys.get(3)),
-                    walletKeys.get(2), tokenid, "info.www.bigtangle.bc", "info.www.bigtangle.bc", aesKey, 678900000,
-                    "");
+                    walletKeys.get(2), tokenid, "info.www.bigtangle.bc", "info.www.bigtangle.bc", aesKey,
+                    BigInteger.valueOf(678900000), "");
             List<ECKey> keys = new ArrayList<ECKey>();
             keys.add(walletKeys.get(3));
             keys.add(preKey);
